testファイルを1つに統合
